VMC/FlightSoftware/status/requirements.txt
loguru==0.5.3
paho-mqtt==1.5.1
Jetson.GPIO==2.0.17
adafruit-circuitpython-register==1.9.8
adafruit-circuitpython-busdevice==5.1.8
adafruit-circuitpython-neopixel-spi==1.0.0